remove unnecessary npm package
