+++ src/components/GalleryImageList.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import ImageList from '@mui/material/ImageList';
 import ImageListItem from '@mui/material/ImageListItem';
 import './GalleryImageList.css';
-import { photos } from './CustomCarousel';
+import { photos } from '../data/photos';
 
 export default function GalleryImageList() {
   const determineColumns = () => {
